+++ packages/astro/test/fixtures/astro-assets-dir/src/content.config.ts
@@ -1,6 +1,8 @@
 import { defineCollection, z } from "astro:content";
+import { glob } from "astro/loaders";
 
 const blogCollection = defineCollection({
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/blog' }),
   schema: ({image}) => z.object({
     title: z.string(),
     cover: image(),