+++ packages/astro/test/fixtures/astro-info-versions/astro.config.mjs
@@ -4,7 +4,10 @@ import node from '@astrojs/node';
 
 // https://astro.build/config
 export default defineConfig({
-	integrations: [react()],
+	integrations: [react(), {
+		name: "manual-integration",
+		hooks: {}
+	}],
 	adapter: node({
 		mode: "standalone"
 	})