+++ packages/astro/test/fixtures/content-collections-mutation/src/content.config.ts
@@ -1,9 +1,10 @@
 // 1. Import utilities from `astro:content`
 import { z, defineCollection } from "astro:content";
+import { glob } from "astro/loaders";
 
 // 2. Define a `type` and `schema` for each collection
 const blogCollection = defineCollection({
-  type: "content", // v2.5.0 and later
+  loader: glob({ pattern: '**/*.{md,mdx}', base: './src/content/blog' }),
   schema: z.object({
     title: z.string(),
     date: z.date(),