+++ packages/astro/test/fixtures/content-intellisense/src/content.config.ts
@@ -2,7 +2,7 @@ import { glob, file } from 'astro/loaders';
 import { defineCollection, z } from 'astro:content';
 
 const blogCC = defineCollection({
-	type: 'content',
+	loader: glob({ pattern: "**/*.{md,mdx,mdoc}", base: "./src/content/blog-cc" }),
 	schema: z.object({
 		title: z.string(),
 		description: z.string().optional(),