+++ packages/astro/test/fixtures/content/astro.config.mjs
@@ -3,8 +3,4 @@ import { defineConfig } from 'astro/config';
 
 export default defineConfig({
   integrations: [mdx()],
-	legacy: {
-		// Enable legacy content collections as we test layout fields
-		collections: true
-	}
 });