+++ packages/integrations/markdoc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @astrojs/markdoc
 
+## 0.15.7
+
+### Patch Changes
+
+- Updated dependencies [[`1e2499e`](https://github.com/withastro/astro/commit/1e2499e8ea83ebfa233a18a7499e1ccf169e56f4)]:
+  - @astrojs/internal-helpers@0.7.3
+  - @astrojs/markdown-remark@6.3.7
+
 ## 0.15.6
 
 ### Patch Changes